Remove a condition for when adding a postgres function
This function was not being added anymore when we moved to django
migrations, because of the condition being removed (Counter is not in
created_models even though it was created). It should be OK to run this
statement on every migration for now, as it adds the function only if it
isn't there yet.