Fix really slow tests
The tests require the whole db to be setup. When we were using south
migrations it took much less time, not because the migrations took less
time, but there seems to have been some mechanism in place that didn't
require the entire db to be rebuilt.

This turns on reuse of the db. It can be rebuilt when required using
`pytest --create-db`.