[django 1.9] Move models imports to where they are needed
Models can't be referenced until after Django is done setting everything
up. This means that either we'll have to change how we import things to
a top package (e.g. all models are available in sentry.models) or just
move the usage of these models to where they actually are needed. That
seems a much simpler option for now.