Clean up migrations and prepare the container model
Migrations needed cleaning up to ensure foreign keys where not
nullable. When adding foreign keys after the fact, they must have a
default value or be not-nullable. A simple solution to that while we
haven't released to production (or staging) is to recreate the
migrations.

This also changes the Container to an extensible model and uses both the
Container and the Substance (which are both extensible) are now referred
to as `archetype` in the Version models. This is so that it's simple to
reuse the code between them cleanly.