Override database user during travis builds
On dev machines, we use passwords to access the dev database. This is to
ensure that people don't accidentally expose sensitive test data in
cases were a dev accidentally opens access to their postgresql and also
happen to keep sensitive test data on their machine.

On travis (and other CI servers) this is not an issue, and for
simplicity we'll just want to run all tests as the travis user. This
patch allows the user to be set with an environment variable so we only
need to set it in travis.yml.