Refactor and change PluginManager
Refactors and changes PluginManager, making it simpler to test and understand.

A major change is that no plugin will be loaded unless it has first been
installed (e.g. with a call to `lims upgrade`).

Another change is that the application will not load by default if there
are plugins that can't load.