Pickup properties of any impl. of a base concept
Before we had to know the ExtensibleType of something to be able to get
it's properties. This is not that useful, since for many purposes we
might want to pickup all volumes, regardless of what ExtensibleType they
were defined on. One example of a usecase for this is to group by
properties in the substance view in the front-end.