Remove sentry-flake8
Sentry-flake8 is catching various errors for py3 which is good, but it's
hardcoded for he sentry codebase and requires certain imports to come
from e.g. sentry.utils.compat (map, zip etc.) which we can't use.

We might not need this at all, since we'll be upgrading to python3 and
killing py2 support (hopefully before v1.0).