Introduce the thread-local context
One of the main goals of Common LIMS is to provide a very simple way to
extend the system. The most basic way to extend the system is in plugin
handlers. The plugin dev should have the power to do everything, but it
should be simple, at least in most cases.

On of the pieces of that design is the thread-local context. Until now
we've just had the "context", which is the same thing but I waited with
implementing this part of it, which is how objects share the context.

The context has:
- the application service, from which all other services can be derived
- the organization
- the current user
(more things may come)

When a handler executes, a context has already been started for the
user. It effectively enhances the business object, because they now have
access to everything the plugin dev requires, without the dev having to
initialize the objects with that data. This means that the user can do
this:

    sample = OurSample(name="testname")
    sample.save()

and it will work, even if they didn't add any other required info, such
as the user or the organization.

A very similar approach *would* have been allowing the user only to
create objects through a factory, e.g.:

    sample = self.OurSample(name="testname")

but since one of the core goals of the plugin system is readability, I
chose the previous approach.

This approach is similar to Django's atomic db transactions, which can be started with
e.g. `with transaction.atomic()`. Django transactions do also use thread local storage to
communicate data between different parts of the code.