Upgrade prettier and eslint
Note that we also stop using eslint-config-sentry-app which means we
have to move some rules into our config instead, as well as some
packages to package.json.

Because we're removing this dependency and for simplicity, I'll follow
this commit up with a `eslint --fix` for everything, which will lead to
a huge one-time change of the codebase, using default prettier
rules.