Update prettier config
Latest version of prettier made us always use parens when calling
lambdas, even with only one argument (or was that some configuration
somewhere that caused it?). We updated all of those, but for some
reason, it's not the setting in my tooling (vim plugin), so we set it
here explicitly.